[SOIN] Déplace la modale d'une sous étape dans un composant dédié
